// src/api/incident-log/models/IncidentSeverity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IncidentLog API
 * Incident reporting and security event management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Severity level of the incident
 * @export
 */
export const IncidentSeverity = {
    Nzka: 'Nízka',
    Stredn: 'Stredná',
    Vysok: 'Vysoká',
    Kritick: 'Kritická'
} as const;
export type IncidentSeverity = typeof IncidentSeverity[keyof typeof IncidentSeverity];


export function IncidentSeverityFromJSON(json: any): IncidentSeverity {
    return IncidentSeverityFromJSONTyped(json, false);
}

export function IncidentSeverityFromJSONTyped(json: any, ignoreDiscriminator: boolean): IncidentSeverity {
    return json as IncidentSeverity;
}

export function IncidentSeverityToJSON(value?: IncidentSeverity | null): any {
    return value as any;
}