// src/api/incident-log/models/IncidentStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IncidentLog API
 * Incident reporting and security event management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Current resolution status of the incident
 * @export
 */
export const IncidentStatus = {
    Nov: 'Nový',
    VRieen: 'V riešení',
    Uzatvoren: 'Uzatvorený'
} as const;
export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export function IncidentStatusFromJSON(json: any): IncidentStatus {
    return IncidentStatusFromJSONTyped(json, false);
}

export function IncidentStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): IncidentStatus {
    return json as IncidentStatus;
}

export function IncidentStatusToJSON(value?: IncidentStatus | null): any {
    return value as any;
}