// src/api/incident-log/models/IncidentType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IncidentLog API
 * Incident reporting and security event management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { IncidentSeverity } from './IncidentSeverity';
import {
    IncidentSeverityFromJSON,
    IncidentSeverityFromJSONTyped,
    IncidentSeverityToJSON,
} from './IncidentSeverity';

/**
 * 
 * @export
 * @interface IncidentType
 */
export interface IncidentType {
    /**
     * 
     * @type {string}
     * @memberof IncidentType
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof IncidentType
     */
    value: string;
    /**
     * 
     * @type {IncidentSeverity}
     * @memberof IncidentType
     */
    typicalSeverity?: IncidentSeverity;
    /**
     * 
     * @type {string}
     * @memberof IncidentType
     */
    description?: string;
}

/**
 * Check if a given object implements the IncidentType interface.
 */
export function instanceOfIncidentType(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "code" in value;
    isInstance = isInstance && "value" in value;

    return isInstance;
}

export function IncidentTypeFromJSON(json: any): IncidentType {
    return IncidentTypeFromJSONTyped(json, false);
}

export function IncidentTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): IncidentType {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'code': json['code'],
        'value': json['value'],
        'typicalSeverity': !exists(json, 'typicalSeverity') ? undefined : IncidentSeverityFromJSON(json['typicalSeverity']),
        'description': !exists(json, 'description') ? undefined : json['description'],
    };
}

export function IncidentTypeToJSON(value?: IncidentType | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'code': value.code,
        'value': value.value,
        'typicalSeverity': IncidentSeverityToJSON(value.typicalSeverity),
        'description': value.description,
    };
}